+++ frontend/.storybook/preview-body.html
@@ -1,2 +1,2 @@
 <div id="modal-root"></div>
-<div id="snack-bar-root"></div>
+<div id="snackbar-root"></div>
