+++ frontend/public/index.html
@@ -9,6 +9,6 @@
   <body>
     <div id="root"></div>
     <div id="modal-root"></div>
-    <div id="snack-bar-root"></div>
+    <div id="snackbar-root"></div>
   </body>
 </html>
