+++ frontend/src/context/snackBar/useSnackBar.ts
@@ -1,5 +1,6 @@
-import { Context } from 'components/@common/SnackBarProvider/SnackBarProvider';
 import { useContext } from 'react';
+
+import { Context } from './SnackBarProvider';
 
 export const useSnackBar = () => {
   const context = useContext(Context);
